code/requirements.txt
Cerberus>=1.3.2
numpy>=1.20.2
matplotlib>=3.3.4
scapy[complete]>=2.4.4
ruamel.yaml>=0.17.2 
docker>=5.0.0
tikzplotlib>=0.9.8
schedule>=1.1.0 